backend/ai/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
boto3==1.26.95
botocore==1.29.95
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
filelock==3.10.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
itypes==1.2.0
jax==0.4.6
Jinja2==3.1.2
jmespath==1.0.1
keras==2.12.0
libclang==15.0.6.1
llvmlite==0.39.1
Markdown==3.4.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==23.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
simplejson==3.18.4
six==1.16.0
sox==1.4.1
sqlparse==0.4.3
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
torch==2.0.0
torchaudio==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
